Move prepareIntegrations helper to utils.
